requirements.txt
asgiref==3.3.4
Django==3.2.3
django-appconf==1.0.4
django-debug-toolbar==3.2.1
django-imagefit==0.6.2
django-js-asset==1.2.2
django-polymorphic==3.0.0
django-request==1.5.6
django-solo==1.1.5
django-tinymce==3.3.0
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
Unidecode==1.2.0
